created a utils file for util functions and shared variables
