+++ local_deploy/get_token.py
@@ -1,16 +1,4 @@
-from square_auth.client_credentials import ClientCredentials
-
-
-def get_token() -> str:
-    client_credentials = ClientCredentials(
-        keycloak_base_url="",
-        buffer=60,
-    )
-    return client_credentials()
-
+from utils import SharedVariables
 
 if __name__ == "__main__":
-    import os
-
-    os.environ["SQUARE_PRIVATE_KEY_FILE"] = os.path.join(os.getcwd(), "private_key.pem")
-    print(get_token())
+    print(SharedVariables.token)
